requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0